split off rust json impl
